+++ pkg/alertmanager/statefulset.go
@@ -110,10 +110,11 @@ func makeStatefulSet(logger *slog.Logger, am *monitoringv1.Alertmanager, config 
 	operator.UpdateObject(
 		statefulset,
 		operator.WithName(prefixedName(am.Name)),
-		operator.WithInputHashAnnotation(inputHash),
 		operator.WithAnnotations(am.GetAnnotations()),
 		operator.WithAnnotations(config.Annotations),
+		operator.WithInputHashAnnotation(inputHash),
 		operator.WithLabels(am.GetLabels()),
+		operator.WithSelectorLabels(spec.Selector),
 		operator.WithLabels(config.Labels),
 		operator.WithManagingOwner(am),
 		operator.WithoutKubectlAnnotations(),
@@ -180,7 +181,7 @@ func makeStatefulSetService(a *monitoringv1.Alertmanager, config Config) *v1.Ser
 
 	svc := &v1.Service{
 		Spec: v1.ServiceSpec{
-			ClusterIP:                "None",
+			ClusterIP:                v1.ClusterIPNone,
 			PublishNotReadyAddresses: true,
 			Ports: []v1.ServicePort{
 				{
@@ -203,7 +204,7 @@ func makeStatefulSetService(a *monitoringv1.Alertmanager, config Config) *v1.Ser
 				},
 			},
 			Selector: map[string]string{
-				"app.kubernetes.io/name": "alertmanager",
+				operator.ApplicationNameLabelKey: applicationNameLabelValue,
 			},
 		},
 	}
@@ -376,7 +377,7 @@ func makeStatefulSetSpec(logger *slog.Logger, a *monitoringv1.Alertmanager, conf
 
 	podAnnotations := map[string]string{}
 	podLabels := map[string]string{
-		"app.kubernetes.io/version": version.String(),
+		operator.ApplicationVersionLabelKey: version.String(),
 	}
 	// In cases where an existing selector label is modified, or a new one is added, new sts cannot match existing pods.
 	// We should try to avoid removing such immutable fields whenever possible since doing
@@ -395,7 +396,7 @@ func makeStatefulSetSpec(logger *slog.Logger, a *monitoringv1.Alertmanager, conf
 		podLabels[k] = v
 	}
 
-	podAnnotations["kubectl.kubernetes.io/default-container"] = "alertmanager"
+	podAnnotations[operator.DefaultContainerAnnotationKey] = "alertmanager"
 
 	var operatorInitContainers []v1.Container
 
@@ -745,11 +746,6 @@ func makeStatefulSetSpec(logger *slog.Logger, a *monitoringv1.Alertmanager, conf
 		return nil, fmt.Errorf("failed to merge containers spec: %w", err)
 	}
 
-	var minReadySeconds int32
-	if a.Spec.MinReadySeconds != nil {
-		minReadySeconds = int32(*a.Spec.MinReadySeconds)
-	}
-
 	operatorInitContainers = append(operatorInitContainers,
 		operator.CreateConfigReloader(
 			"init-config-reloader",
@@ -774,7 +770,7 @@ func makeStatefulSetSpec(logger *slog.Logger, a *monitoringv1.Alertmanager, conf
 	spec := appsv1.StatefulSetSpec{
 		ServiceName:     getServiceName(a),
 		Replicas:        a.Spec.Replicas,
-		MinReadySeconds: minReadySeconds,
+		MinReadySeconds: ptr.Deref(a.Spec.MinReadySeconds, 0),
 		// PodManagementPolicy is set to Parallel to mitigate issues in kubernetes: https://github.com/kubernetes/kubernetes/issues/60164
 		// This is also mentioned as one of limitations of StatefulSets: https://kubernetes.io/docs/concepts/workloads/controllers/statefulset/#limitations
 		PodManagementPolicy: appsv1.ParallelPodManagement,
@@ -804,6 +800,7 @@ func makeStatefulSetSpec(logger *slog.Logger, a *monitoringv1.Alertmanager, conf
 				TopologySpreadConstraints:     a.Spec.TopologySpreadConstraints,
 				HostAliases:                   operator.MakeHostAliases(a.Spec.HostAliases),
 				EnableServiceLinks:            a.Spec.EnableServiceLinks,
+				HostUsers:                     a.Spec.HostUsers,
 			},
 		},
 	}