+++ pkg/apis/monitoring/v1/types.go
@@ -988,6 +988,8 @@ type NativeHistogramConfig struct {
 	// Whether to scrape a classic histogram that is also exposed as a native histogram.
 	// It requires Prometheus >= v2.45.0.
 	//
+	// Notice: `scrapeClassicHistograms` corresponds to the `always_scrape_classic_histograms` field in the Prometheus configuration.
+	//
 	// +optional
 	ScrapeClassicHistograms *bool `json:"scrapeClassicHistograms,omitempty"`
 
@@ -1019,6 +1021,16 @@ const (
 	SelectorMechanismRelabel SelectorMechanism = "RelabelConfig"
 	SelectorMechanismRole    SelectorMechanism = "RoleSelector"
 )
+
+// ConfigResourceStatus is the most recent observed status of the Configuration Resource (ServiceMonitor, PodMonitor and Probes). Read-only.
+// More info:
+// https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#spec-and-status
+// +k8s:openapi-gen=true
+type ConfigResourceStatus struct {
+	// The list of workload resources (Prometheus or PrometheusAgent) which select the configuration resource.
+	// +optional
+	Bindings []WorkloadBinding `json:"bindings,omitempty"`
+}
 
 // WorkloadBinding is a link between a configuration resource and a workload resource.
 // +k8s:openapi-gen=true